backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
beanie==1.26.0
motor==3.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
openai>=1.52.0
python-dotenv==1.0.1
pypdf==4.3.1
pymupdf==1.27.2.2
python-docx==1.1.2
matplotlib==3.9.2
numpy==2.1.1
